requirements.txt
# ML
mpi4py
tensorflow
torch
stable-baselines3 
torchtyping

# tensorboard==2.12.0 
ribs 
dm_alchemy @ git+https://github.com/robbycostales/dm_alchemy.git
xminigrid @ git+https://github.com/robbycostales/xland-minigrid.git

# Math
numpy #==1.22.1
scipy
scikit-learn
einops

# Misc from DSAGE
gin-config

# Jupyter lab
jupyterlab
jupyterlab_widgets
ipywidgets
ipympl

# Others
gymnasium
matplotlib
seaborn
tqdm
psutil
numba
ruff  # code quality

# Debugging
pyinstrument
pytest
pytest-cov

# Dask
dask
distributed
click
bokeh
jupyter-server-proxy

# Maze
gym-minigrid
scikit-image

# QD+MetaRL
gym==0.25.1 # ==0.18.0
chardet #==5.1.0
xxhash #==3.3.0
moviepy #==1.0.3

# Logging
wandb #==0.15.8
guppy3 #==3.1.4.post1
objgraph #==3.6.0
loguru

# DCD CarRacing environment
pygame
pyglet==1.5.27  # Older version necessary for pyglet.gl classes
# 2.0.17 gave: "NameError: name 'glPushMatrix' is not defined" when rendering for racing
box2d-py #==2.3.5
# gym[box2d]==0.25.1
# Box2D==2.3.2
box2d-kengz #==2.3.3
pyvirtualdisplay #==3.0
geopandas #==0.9.0  # Installs click 8.1.6, which conflicts with dask version.

# DeepMind Alchemy
absl-py #==1.4.0
dataclasses #==0.6
dm-env #==1.6
dm-env-rpc #==1.1.5   #>=1.0.4
dm-tree #==0.1.8
docker #==6.1.3
grpcio #==1.57.0
portpicker #==1.6.0
frozendict #==2.3.8
vpython # For visualizer
mayavi #==4.8.1  # For visualizer

# Hydra
hydra-core
omegaconf
hydra-submitit-launcher